+++ module-newsfeed-service/src/main/java/com/shop/newsfeedservice/common/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.shop.newsfeedservice.config;
+package com.shop.newsfeedservice.common.config;
 
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
