+++ module-newsfeed-service/src/main/java/com/shop/newsfeedservice/common/exception/BaseException.java
@@ -1,5 +1,6 @@
-package com.shop.newsfeedservice.exception;
+package com.shop.newsfeedservice.common.exception;
 
+import com.shop.newsfeedservice.common.response.ErrorCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
