+++ module-newsfeed-service/src/main/java/com/shop/newsfeedservice/newsfeed/entity/Newsfeed.java
@@ -1,5 +1,6 @@
-package com.shop.newsfeedservice.domain;
+package com.shop.newsfeedservice.newsfeed.entity;
 
+import com.shop.newsfeedservice.common.entity.BaseTimeEntity;
 import jakarta.persistence.*;
 import lombok.*;
